web_scraping/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
courlan==1.3.2
crawl4ai==0.6.3
cryptography==45.0.5
cssselect==1.3.0
dateparser==1.2.2
distro==1.9.0
dnspython==2.7.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.116.0
fasttext-predict==0.9.2.4
ffmpy==0.6.0
filelock==3.18.0
flask==3.1.1
flask-babel==4.0.0
frozenlist==1.7.0
fsspec==2025.5.1
gradio==5.35.0
gradio-client==1.10.4
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
httpx-socks==0.10.0
huggingface-hub==0.33.2
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
justext==3.0.2
litellm==1.74.0.post1
lxml==5.4.0
lxml-html-clean==0.4.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
multidict==6.6.3
nltk==3.9.1
numpy==2.3.1
openai==1.93.1
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==10.4.0
playwright==1.53.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.10.1
pydub==0.25.1
pyee==13.0.0
pygments==2.19.1
pymongo==4.13.2
pyopenssl==25.1.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-socks==2.7.1
pytz==2025.2
pyyaml==6.0.2
rank-bm25==0.2.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.2
safehttpx==0.1.6
semantic-version==2.10.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
starlette==0.46.2
tavily-python==0.7.9
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tld==0.13.1
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
trafilatura==2.0.0
typer==0.16.0
typer-slim==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
websockets==15.0.1
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
